refactor: 폴더 구조 변경 (#6) - Input 컴포넌트 common 폴더로 이동 - fileds 폴더 따로 분리
