+++ src/components/fields/file-field.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import * as React from "react";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
+import { Input } from "@/components/common/Input";
+import { Button } from "@/components/ui/button_01";
 import { cn } from "@/lib/utils";
 
 export type FileFieldProps = {
