+++ src/components/fields/select-field.tsx
@@ -22,9 +22,9 @@ export type SelectFieldProps = {
 };
 
 const sizeMap = {
-  sm: "h-10 px-4 text-sm",
+  sm: "h-10 px-4 ",
   md: "h-12 px-4",
-  lg: "h-14 px-4 text-base",
+  lg: "h-14 px-4 ",
 };
 
 export const SelectField = React.forwardRef<HTMLDivElement, SelectFieldProps>(
@@ -47,11 +47,21 @@ export const SelectField = React.forwardRef<HTMLDivElement, SelectFieldProps>(
           <SelectTrigger
             aria-invalid={state === "invalid" || undefined}
             className={cn(
-              "w-full rounded-md border bg-transparent px-3 text-left shadow-xs outline-none transition-[box-shadow,border-color] focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0",
-              "border-input",
+              // 레이아웃 & 폭
+              "w-full text-left",
+              // 패딩 & 사이즈 매핑
+              sizeMap[size],
+              // 배경 & 테두리 & 라운드
+              "bg-transparent border border-input rounded-xl",
+              // 그림자
+              "shadow-xs",
+              // 포커스
+              "outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-0",
+              // 전환 효과
+              "transition-[box-shadow,border-color]",
+              // 상태 경계선
               state === "invalid" && "border-red-500",
               state === "success" && "border-purple-500",
-              sizeMap[size],
               className,
             )}
           >
